--- .bowerrc
@@ -1,3 +1,0 @@
-{
-    "directory": "bower_components"
-}
