--- .jshintrc
@@ -1,37 +1,0 @@
-{
-  "node": true,
-  "browser": true,
-  "esnext": true,
-  "bitwise": true,
-  "camelcase": true,
-  "curly": true,
-  "eqeqeq": true,
-  "immed": true,
-  "indent": 2,
-  "latedef": false,
-  "newcap": true,
-  "noarg": true,
-  "quotmark": "single",
-  "regexp": true,
-  "smarttabs": true,
-  "strict": true,
-  "trailing": true,
-  "undef": true,
-  "validthis": true,
-  "-W058": true,
-  "predef": [
-    "$",
-    "jQuery",
-    "before",
-    "beforeEach",
-    "define",
-    "describe",
-    "describeComponent",
-    "describeMixin",
-    "expect",
-    "it",
-    "requirejs",
-    "setupComponent",
-    "spyOnEvent"
-  ]
-}
