--- bower.json
@@ -1,21 +1,0 @@
-{
-  "name": "flight-with-resources",
-  "version": "1.0.0",
-  "main": "lib/with-resources.js",
-  "dependencies": {
-    "flight": "^1.2.0"
-  },
-  "devDependencies": {
-    "jasmine-flight": "^2.2.0"
-  },
-  "ignore": [
-    ".gitignore",
-    ".gitattributes",
-    ".travis.yml",
-    "CONTRIBUTING.md",
-    "CHANGELOG.md",
-    "test",
-    "package.json",
-    "karma.conf.js"
-  ]
-}
