+++ package.json
@@ -1,26 +1,48 @@
 {
   "name": "flight-with-resources",
-  "version": "1.0.0",
-  "devDependencies": {
-    "grunt": "~0.4.5",
-    "grunt-bump": "^0.0.15",
-    "karma": "~0.12.6",
-    "karma-cli": "0.0.4",
-    "karma-jasmine": "~0.1.0",
-    "karma-requirejs": "~0.2.2",
-    "karma-chrome-launcher": "*",
-    "karma-firefox-launcher": "*",
-    "matchdep": "^0.3.0",
-    "requirejs": "^2.1.15"
-  },
-  "scripts": {
-    "test": "karma start --single-run --reporters dots --browsers Firefox",
-    "watch-test": "karma start"
-  },
   "description": "A Flight mixin for sharing named resources between components.",
-  "main": "lib/with-resources.js",
-  "directories": {
-    "test": "test"
+  "main": "dist/flight-with-resources.js",
+  "scripts": {
+    "build": "rm -rf ./dist && NODE_ENV=publish webpack --config config/webpack.config.publish.js --sort-assets-by --progress",
+    "lint": "eslint config src",
+    "lint:fix": "eslint --fix config src",
+    "prepublish": "npm run build",
+    "specs": "NODE_ENV=test karma start config/karma.config.js",
+    "specs:watch": "npm run specs -- --no-single-run",
+    "test": "npm run specs && npm run lint",
+    "semantic-release": "semantic-release pre && npm publish && semantic-release post"
+  },
+  "devDependencies": {
+    "babel-core": "^5.8.24",
+    "babel-eslint": "^4.1.1",
+    "babel-loader": "^5.3.2",
+    "babel-plugin-typecheck": "^1.2.0",
+    "babel-runtime": "^5.8.20",
+    "chai": "^3.2.0",
+    "eslint": "^1.3.1",
+    "eslint-config-standard": "^4.3.1",
+    "eslint-config-standard-react": "^1.0.4",
+    "eslint-plugin-react": "^3.3.1",
+    "eslint-plugin-standard": "^1.3.0",
+    "flightjs": "^1.5.1",
+    "immutable": "^3.7.5",
+    "jasmine-core": "^2.3.4",
+    "jquery": "^2.1.4",
+    "karma": "^0.13.9",
+    "karma-chrome-launcher": "^0.2.0",
+    "karma-cli": "^0.1.0",
+    "karma-firefox-launcher": "^0.1.6",
+    "karma-jasmine": "^0.3.6",
+    "karma-mocha": "^0.2.0",
+    "karma-sourcemap-loader": "^0.3.5",
+    "karma-webpack": "^1.7.0",
+    "mocha": "^2.3.2",
+    "object-assign": "^4.0.1",
+    "semantic-release": "^4.3.5",
+    "webpack": "^1.12.1"
+  },
+  "peerDependencies": {
+    "flightjs": "^1.5.1"
   },
   "repository": {
     "type": "git",
